+++ src/uncms/testhelpers/models.py
@@ -1,0 +1,5 @@
+from uncms.pages.models import ContentBase
+
+
+class EmptyTestPage(ContentBase):
+    pass